app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
openai==1.51.0
databricks-sdk==0.34.0
databricks-sql-connector==3.4.0
